pcml/functions/embed_frames/requirements.txt
google-cloud-firestore==0.31.0
tensorflow==1.13.1
numpy